fix: Python 3.14 compatibility
Python 3.14 introduced changes to annotations, which broke msgspec
structs; unfortunately, msgspec is no longer maintained.

A workaround is to import annotations from __future__. This draft commit
adds this import to all Python files, but we really only need it for
some subset of these files (not yet sure which subset).

* issue: https://github.com/jcrist/msgspec/issues/847
* example workaround:
  https://github.com/return42/searxng/commit/3963c2371a3a33728771ebef0d173b3d47fc4e11