fix: msgspec Python 3.14 compatibility (#6749)
Python 3.14 introduced changes to annotations, which broke `msgspec`
structs; unfortunately, `msgspec` is no longer actively maintained.

While we search for an alternative or long-term solution, a short-term
workaround is to import annotations from `__future__`.

* issue: https://github.com/jcrist/msgspec/issues/847
* example workaround:
https://github.com/return42/searxng/commit/3963c2371a3a33728771ebef0d173b3d47fc4e11

This PR adds the import to more modules than needed; not sure what the
minimal set is.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dylan Madisetti <dylan@marimo.io>