fix empty submits and output area (#6794)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->

There was an issue on browser refresh where `Fix with AI` wouldn't do
anything, it's because input is empty. We now use the `complete` API to
force set the input state and trigger completion.

Also for outputArea above, we put the banner at the top of the cell
instead of at the bottom.

![CleanShot 2025-10-15 at 22 33
26](https://github.com/user-attachments/assets/1fe5cfbe-fbc8-418d-977f-c2d10ffec7a3)

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.