fix: duckdb interrupt (#6806)
This change adds more graceful interrupts of DuckDB queries executed
through `mo.sql()`.

Interrupting a connection with a `SIGINT` can apparently leave the
connection in a bad state. This change instead interrupts the connection
with its `interrupt()` method. To enable this, the `ExecutionContext` is
augmented with an additional property in which the runtime can store the
active duckdb connection.

This is difficult to test programmatically. Tested manually by
interrupting the query in
https://github.com/marimo-team/marimo/issues/6787. Basic unit tests are
included to ensure the interrupt handler does the right thing.