docs: dataframe docs (#6812)
I figured I'd add some extra context on the dataframe settings when I
noticed that the dataframe docs were broken 🫨 !


https://github.com/user-attachments/assets/ed1c3b2d-20ad-427f-847b-15cc25406cc8

This PR fixes the issue, we were missing some closing `///` that would
hide whole portions of the docs as a result.

It also mentions a user setting that is relevant to this portion of the
docs.

---------

Co-authored-by: Myles Scolnick <myles@marimo.io>