break: update markdown output to include 'contents' class (#6804)
`display: contents` causes the element to not generate a box in the
layout tree, so wrapped markdown class does not effect the layout of the
child. this can be useful if the child is something different than a
span (e.g. table)

This may be a breaking change for some layouts, but I could not produce
a way that it would break in an unexpected way.

Fixes https://github.com/marimo-team/marimo/issues/6802