deps: upgrade form-data to secure version (#6888)
Dependabot could not upgrade form-data along because of multiple major
versions, so this overrides form-data in the package json to be on a
version without a security warning attached.

I looked through some of the deps that pull in the old form-data (v2)
and it looked to be non-breaking changes or code paths that are not
used.