Document `AppMeta` and make it public (#6939)
This PR was made apon reflecting [this message on
Discord](https://discord.com/channels/1059888774789730424/1179936876103225354/1432299089873211443).

The user in this case was trying to figure out whether or not there was
a present mode whenever you access the app meta information. The
documentation doesn't actually confirm if this is the case. It doesn't
list all the values that the run mode can represent.

Hence, I figured adding the `AppMeta` class to our documentation. That
way a user could at least confirm themselves if it is/not supported.