improvement: support binning point selection in altair (#6983)
Fixes #6829 

This adds support for binning + chart selection. This also works for
binning in the legend as well.

One thing I could not get working in the visual highlighting for
interval selection + binning. The data filtering does work, but not the
highlighting, so i log a warning to the user.

<img width="1183" height="1104" alt="Screenshot 2025-10-29 at 10 20
48 AM"
src="https://github.com/user-attachments/assets/3973e788-64c7-4879-8a4c-54f19384e06b"
/>