fix: Prevent matching by cell_id (#7072)
## 📝 Summary

When code_hash was empty, we defaulted to matching session outputs by
position. This is less than ideal since the outputs may be very very
different.

Will followup with a PR that enforces no empty `code_hash`, mitigating
some of the cases where we could output something to the user.