improve: do not trigger UI element warnings for state updates (#7102)
Right now the runtime eagerly updates states value when a state update
is registered. This should be unnecessary (but maybe was needed
for file watching element, which is when this code was introduced),
and it hits a code path for UI elements that triggers a very loud error.
Instead of fixing the underlying issue, this PR prevents the
code path from being triggered and prints a gentler warning.

Related: #7060