Fix: state resolution (#7110)
This PR prunes logic in state resolution which eagerly marked cells as
stale and enqueued a stale cell request whenever a setter was called; in
the main thread, running cells made stale by a state update is handled
by the main cell execution loop.

Eager execution of state updates only needs to happen when running state
updates in a `mo.Thread` object (either manually or via the mo.watch
APIs). This PR updates conditions how state resolution happens based on
whether the current thread is a `mo.Thread`.

Additionally, instead of relying on the state registry to find bound
names, which does not work when a state update is called in the cell
that defined it, we just scan globals to find cells that have the state
object in their refs (similar to what `CellRunner` does).

Closes #7060
Closes https://github.com/marimo-team/marimo/issues/6847
Closes https://github.com/marimo-team/marimo/issues/5311
Closes https://github.com/marimo-team/marimo/issues/5139