fix: the default rendering of ibis tables when interactive should still be lazy (#7163)
Ibis is naturally lazy and even in "interactive" mode, only render the
first 10 items. For our opinionated DF renders, we were loading the full
table into memory to display with pagination, but this could load large
datasets that were not intended. So instead we show a lazy-table which
only shows the first 10 items as well (but using our opinionated table).

If users do want to load the full ibis table, they can call `.execute()`
or pass the table into `mo.ui.table(df)`