feat: Contextually change stat color (#7175)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->

With this PR it possible to manually swap the color of the triangle in
the `mo.stat` UI element. Previously the color was always green for
increases and always red for decreases, but sometimes a decrease is
"good" and one may want to display a green downwards facing triangle.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

The `stat` function now has a `color` argument that accepts the literals
"positive" (green) or "negative" (red) to allow the user to set the
color manually. If not specified the color will be as it is now, green
for increase and red for decrease.

## 📋 Checklist

- [X] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [X] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
https://discord.com/channels/1059888774789730424/1059891311190229082/1438631478178353232
- [X] I have added tests for the changes made.
- [X] I have run the code and verified that it works as expected.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>