Make workspace refresh button accessible (#7213)
## 📝 Summary

Currently it isn't possible to focus the refresh button on the Marimo
local server's home page using a keyboard. This change makes the button
more accessible by using an actual button tag.

## 🔍 Description of Changes

I changed just the one component from an icon to a button that wraps the
icon. This causes the button to inherit the usual browser accessibility
features.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.