api: remove stream argument for built-in AI models (#7222)
This PR simplifies the API for the built-in chat models, removing the
stream argument and instead always streaming responses (which is a much
better experience than synchronous responses). All built-in models
already support streaming, so this should just work.

(This does not change the API for custom models, only built-in models)

Removing the argument itself is not breaking change because streaming
support (added in #7187) has not been released. However this does change
the model return type, which is technically breaking (but we do not
anticipate users calling models directly, and instead expect users to
use models in `mo.ui.chat()`).