remove snippet submission (#7404)
Fixes #7397

This PR removes the snippet form. This feature is not useful/used at
all, and can pose a privacy risk for those who don't want to submit
snippets. We also now support custom snippets
(https://docs.marimo.io/guides/configuration/snippets/), and it is
easier to configure this outside of the editor for now.