Unify handling of widget state/buffers/bufferPaths (#7441)
This keeps the `state/buffer/bufferPaths` in sync when sending to/from
the backend for the widget. This changes some typings and adds a few
more tests. This is an improvement, but not the final state as we can
hopefully simplify further. This does fix a few errors when sending back
binary data.

Prior work done by @manzt in
https://github.com/marimo-team/marimo/pull/7441/commits/9426b131008d827f207f85f53274d7f3ee71dac3

---------

Co-authored-by: Trevor Manz <trevor.j.manz@gmail.com>