feat: CLI flag to configure whether a notebook is trusted (#7556)
This PR adds a new flag to `marimo edit` and `marimo run`: `--trusted`.
This flag controls whether marimo is run in a Docker container.

* `marimo edit --trusted ...`: does not run in a container
* `marimo edit --untrusted ...`: runs in a Docker container
* `marimo edit <NOTEBOOK>`: prompts whether to run in Docker if
`NOTEBOOK` is determined to be hosted remotely; this behavior is the
same as it was before this change.

This enables users to run trusted notebooks that are hosted remotely on
the host machine, without having to respond interactive prompts. For
example, `uvx marimo edit --trusted --sandbox URL`