internal: session extensions (#7542)
This refactors the `Session` class into extensions that attach/detach
and can optionally listen to session events. This is the start of the
refactor, but by the end, it should essentially just be an event bus.
This really reduces the size of the session (and conditional logic)
making it more composable and re-usable for different use cases
(marimo-lsp) and different condition (read/edit, autoinstantiate/not)