Revamp Ollama models list (#7595)
This is my personal short-list where I err on the side of lighter
compute.

The main important thing s that this list is just very short now. Folks
are bound to toy around with new models once they are released. So the
main edit here is to prevent unneeded scrolling when people want to try
out something new.