improvement: run auto-instaniate on the backend for run mode (#7603)
This PR makes the `/instantiate` call happen on the backend for run
mode. The backend has all the information to start the session and does
not need to have the frontend call it. Doing this allows us to:
1. require `edit` perms to call `/instantiate`, and also:
2. optionally accept `codes: dict[CellId, str]` to instantiate. 

This lets us lazily call instantiate so we can allow edits or add/remove
cells, and then lazily start a kernel by using the new local state.